Fix benchmark script to work with poetry:
When initially refactoring to use poetry, the Dockerfile used in
the benchmark task stored the virtual env directory within the
package working directory.

However, the github actions benchmark script maps the runner's working
copy of the ucc directory to the dockerfile run of the benchmark script.
This shadows the /ucc/.venv directory.

This fix switches to store the virtual environment directory outside
the package working directory to avoid this shadowing.